type declarations: explicit imports and separate files for esm and commonjs
based on the typescript handbook, I understand that explicit file
extensions are the recommended approach, whether the project uses
esm or commonjs. This appears to be _mandatory_ with the recommended
setting "moduleResolution": "nodenext" and "module": "nodenext" -
otherwise tsc will detect potential errors when in checks the
resolvability of imports.

I'm _not_ sure if the separate declaration files for esm (.d.mts) and
commonjs (.d.cjs) are necessary, but this fixed an error that was
reported by the attw tool. The (.d.ts) files are probably reduntant /
equivalent to their .d.mts copies.